LPS-179353 Extract ScreenNavigationEntry methods from RemoteCommerceTaxMethodConfigurationScreenNavigationCategory to RemoteCommerceTaxMethodConfigurationScreenNavigationEntry, extends RemoteCommerceTaxMethodConfigurationScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
